JDBC

+++ src/Training/Ex14/Model/GoodStudent.java
@@ -1,20 +1,20 @@
 package src.Training.Ex14.Model;
 
 public class GoodStudent extends Student {
-    private float gpa;
+    private double gpa;
     private String bestRewardName;
 
-    public GoodStudent(String fullName, String dateOfBirth, String sex, String phoneNumber, String universityName, String gradeLevel, float gpa, String bestRewardName) {
+    public GoodStudent(String fullName, String dateOfBirth, String sex, String phoneNumber, String universityName, String gradeLevel, double gpa, String bestRewardName) {
         super(fullName, dateOfBirth, sex, phoneNumber, universityName, gradeLevel);
         this.gpa = gpa;
         this.bestRewardName = bestRewardName;
     }
 
-    public float getGpa() {
+    public double getGpa() {
         return gpa;
     }
 
-    public void setGpa(float gpa) {
+    public void setGpa(double gpa) {
         this.gpa = gpa;
     }
