+++ src/Training/Ex14/Model/NormalStudent.java
@@ -1,28 +1,28 @@
 package src.Training.Ex14.Model;
 
 public class NormalStudent extends Student {
-    private int englishScore;
-    private int entryTestScore;
+    private double englishScore;
+    private double entryTestScore;
 
-    public NormalStudent(String fullName, String dateOfBirth, String sex, String phoneNumber, String universityName, String gradeLevel, int englishScore, int entryTestScore) {
+    public NormalStudent(String fullName, String dateOfBirth, String sex, String phoneNumber, String universityName, String gradeLevel, double englishScore, double entryTestScore) {
         super(fullName, dateOfBirth, sex, phoneNumber, universityName, gradeLevel);
         this.englishScore = englishScore;
         this.entryTestScore = entryTestScore;
     }
 
-    public int getEnglishScore() {
+    public double getEnglishScore() {
         return englishScore;
     }
 
-    public void setEnglishScore(int englishScore) {
+    public void setEnglishScore(double englishScore) {
         this.englishScore = englishScore;
     }
 
-    public int getEntryTestScore() {
+    public double getEntryTestScore() {
         return entryTestScore;
     }
 
-    public void setEntryTestScore(int entryTestScore) {
+    public void setEntryTestScore(double entryTestScore) {
         this.entryTestScore = entryTestScore;
     }
